+++ contracts/token/psp1155/src/extensions/metadata.rs
@@ -1,0 +1,31 @@
+use crate::traits::Id;
+/// Metadata for PSP1155
+use brush::{
+    declare_storage_trait,
+    traits::InkStorage,
+};
+use ink_storage::traits::SpreadLayout;
+pub use psp1155_derive::{
+    PSP1155MetadataStorage,
+    PSP1155Storage,
+};
+
+#[cfg(feature = "std")]
+use ink_storage::traits::StorageLayout;
+
+#[derive(Default, Debug, SpreadLayout)]
+#[cfg_attr(feature = "std", derive(StorageLayout))]
+pub struct PSP1155MetadataData {
+    pub uri: Option<String>,
+}
+
+declare_storage_trait!(PSP1155MetadataStorage, PSP1155MetadataData);
+
+#[brush::trait_definition]
+pub trait PSP1155Metadata: PSP1155MetadataStorage {
+    /// Returns the uri for token type of id.
+    #[ink(message)]
+    fn uri(&self, _id: Id) -> Option<String> {
+        self.get().uri.clone()
+    }
+}
