+++ contracts/token/psp1155/src/extensions/mod.rs
@@ -1,2 +1,3 @@
 pub mod burnable;
+pub mod metadata;
 pub mod mintable;
