+++ contracts/token/psp1155/tests/tests.rs
@@ -10,6 +10,7 @@ mod tests {
     use psp1155::{
         extensions::{
             burnable::*,
+            metadata::*,
             mintable::*,
         },
         traits::*,
